Implement Zorro SHUTUP function
Turns out it is not required to use a pin for SLAVE.OE generation,
as only one term is used. That lets us reuse Pin 21 for something
else, i.e. the Zorro SHUTUP function.
If the 1GB Zorro III memory range is completely filled up (for
example with two 512MB memory cards), there is no more space for
an A4091. This leads to other cards (even Zorro II I/O cards) to
not be configured anymore if they are in the slot behind the
problematic board, as Autoconfig is blocked at this point.
This happens because CFGOUT is never activated and so the next
Zorro card never becomes active.

Found by dorken @ a1k.org